+++ sandbox/settings/base.py
@@ -139,6 +139,8 @@ TEMPLATES = [
 ]
 
 INSTALLED_APPS = [
+    "dal",
+    "dal_select2",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",