+++ tests/050_views/054_tag.py
@@ -38,9 +38,6 @@ def test_tag_view_index(db, admin_client, client, enable_preview, settings):
     settings.LANGUAGE_CODE = "en"
 
     utc = ZoneInfo("UTC")
-    yesterday = datetime.datetime(2012, 10, 14, 10, 0).replace(tzinfo=utc)
-    tomorrow = datetime.datetime(2012, 10, 16, 10, 0).replace(tzinfo=utc)
-    past_hour = datetime.datetime(2012, 10, 15, 9, 00).replace(tzinfo=utc)
     next_hour = datetime.datetime(2012, 10, 15, 11, 00).replace(tzinfo=utc)
 
     picsou = AuthorFactory(first_name="Picsou", last_name="McDuck")
@@ -135,8 +132,6 @@ def test_tag_view_index(db, admin_client, client, enable_preview, settings):
         }
         for item in container.cssselect("a")
     ]
-    # print()
-    # print(json.dumps(tags, indent=4))
     assert tags == [
         {
             "name": "French only",
@@ -190,4 +185,170 @@ def test_tag_view_index(db, admin_client, client, enable_preview, settings):
             "count": 3,
             "url": "/en/tags/science/"
         }
+    ]
+
+
+@freeze_time("2012-10-15 10:00:00")
+def test_tag_view_detail(db, admin_client, client, enable_preview, settings):
+    """
+    Tag detail should list its related article respecting criterias and article list
+    pagination should be triggered when there is enough articles.
+    """
+    # NOTE: A test playing with language and view requests must enforce default
+    # language since LANGUAGE_CODE may be altered between two tests.
+    settings.LANGUAGE_CODE = "en"
+
+    utc = ZoneInfo("UTC")
+    next_hour = datetime.datetime(2012, 10, 15, 11, 00).replace(tzinfo=utc)
+
+    picsou = AuthorFactory(first_name="Picsou", last_name="McDuck")
+
+    cat_1 = CategoryFactory(title="cat_1")
+
+    science = TagFactory(name="Science", slug="science")
+    sausage = TagFactory(name="Sausage", slug="sausage")
+    TagFactory(name="Not used", slug="not-used")
+
+    ArticleFactory(title="Foo", fill_tags=[science])
+    ArticleFactory(title="Bar",fill_tags=[science])
+    # Unavailable articles to anonymous
+    ArticleFactory(
+        title="France",
+        fill_tags=[science],
+        language="fr",
+    )
+    # Unavailable articles to anonymous
+    ArticleFactory(
+        title="Ping",
+        fill_tags=[science],
+        status=STATUS_DRAFT,
+    )
+    ArticleFactory(
+        title="Pong",
+        fill_tags=[science],
+        private=True,
+    )
+    ArticleFactory(
+        title="Pang",
+        fill_tags=[science],
+        publish_date=next_hour.date(),
+        publish_time=next_hour.time(),
+    )
+    ArticleFactory(
+        title="Nope",
+        fill_tags=[sausage],
+    )
+
+    # Get tag index url for lambda user with defaut language 'en'
+    url = reverse("lotus:tag-detail", kwargs={"tag": science.slug})
+    # Get page HTML
+    response = client.get(url)
+    assert response.status_code == 200
+    # Parse HTML response  and build a summary from found items so we can easily make
+    # assertion
+    dom = html_pyquery(response)
+    items = dom.find("#lotus-content .articles .article")
+    content = []
+    for item in items:
+        title = item.cssselect(".title")[0].text
+        content.append(title)
+
+    assert content == ["Bar", "Foo"]
+
+    # No pagination HTML is displayed when limit have not been reached
+    pagination = dom.find("#lotus-content .articles .pagination")
+    assert len(pagination) == 0
+
+    # Admin in preview mode should be able to see every article related to the tag
+    enable_preview(admin_client)
+    response = admin_client.get(url)
+    assert response.status_code == 200
+    dom = html_pyquery(response)
+    items = dom.find("#lotus-content .articles .article")
+    content = []
+    for item in items:
+        title = item.cssselect(".title")[0].text
+        content.append(title)
+
+    assert content == ["Pang", "Bar", "Foo", "Ping", "Pong"]
+
+    # Create some other articles to trigger pagination limit
+    ArticleFactory.create_batch(
+        settings.LOTUS_ARTICLE_PAGINATION,
+        fill_tags=[science],
+    )
+    response = client.get(url)
+    assert response.status_code == 200
+    dom = html_pyquery(response)
+    pagination = dom.find("#lotus-content .articles .pagination")
+    assert len(pagination) == 1
+
+
+def test_tag_view_autocomplete(db, admin_client, client, enable_preview, settings):
+    """
+    Autocomplete view should be only available for admins, returning list in JSON to
+    the requests.
+    """
+    user = AuthorFactory()
+
+    science = TagFactory(name="Science theme", slug="science")
+    sausage = TagFactory(name="Sausage food", slug="sausage")
+    game = TagFactory(name="Game theme", slug="game")
+    ping = TagFactory(name="Ping", slug="ping")
+    pingpong = TagFactory(name="Ping pong", slug="ping-pong")
+    music = TagFactory(name="Music theme", slug="music")
+    french_only = TagFactory(name="French theme", slug="french")
+    not_available = TagFactory(name="Not available", slug="not-available")
+
+    url = reverse("lotus:tag-autocomplete")
+
+    # Anonymous can not reach the view
+    response = client.get(url)
+    assert response.status_code == 302
+
+    # Simple logged in user can not reach the view
+    client.force_login(user)
+    response = client.get(url)
+    assert response.status_code == 403
+
+    # Simple get request return every tags (with possible pagination)
+    response = admin_client.get(url)
+    assert response.status_code == 200
+    json_data = response.json()
+    tags = [item["text"] for item in json_data["results"]]
+    assert tags == [
+        "French theme",
+        "Game theme",
+        "Music theme",
+        "Not available",
+        "Ping",
+        "Ping pong",
+        "Sausage food",
+        "Science theme",
+    ]
+    assert json_data["pagination"] == {"more": False}
+
+    # POST request is forbidden
+    response = admin_client.post(url, {"q": "Ping"}, HTTP_ACCEPT="application/json")
+    assert response.status_code == 400
+
+    # Demonstrate queryset match which is basically a "startswith"
+    response = admin_client.get(url, {"q": "S"}, HTTP_ACCEPT="application/json")
+    assert [item["text"] for item in response.json()["results"]] == [
+        "Sausage food",
+        "Science theme",
+    ]
+
+    response = admin_client.get(url, {"q": "Niet"}, HTTP_ACCEPT="application/json")
+    assert [item["text"] for item in response.json()["results"]] == []
+
+    response = admin_client.get(url, {"q": "theme"}, HTTP_ACCEPT="application/json")
+    assert [item["text"] for item in response.json()["results"]] == []
+
+    response = admin_client.get(url, {"q": "Scie"}, HTTP_ACCEPT="application/json")
+    assert [item["text"] for item in response.json()["results"]] == ["Science theme"]
+
+    response = admin_client.get(url, {"q": "Ping"}, HTTP_ACCEPT="application/json")
+    assert [item["text"] for item in response.json()["results"]] == [
+        "Ping", "Ping pong"
     ]