+++ docs/history.rst
@@ -16,9 +16,12 @@ Version 0.5.1 - Unreleased
   requirements "dev";
 * Added 'Translate' link to Article detail page along the 'Edit' link;
 * Added Article tags feature with ``django-taggit``;
+* Added ``django-autocomplete-light`` for a nice widget on Article 'tags' field in
+  admin;
 * Changed models ``get_absolute_url`` method to use translate_url instead of
   ``translation_activate``;
-* Added ``lookups.LookupBuilder`` an abstraction to make complex lookups for publication/language criterias for Article and Category;
+* Added ``lookups.LookupBuilder`` an abstraction to make complex lookups for
+  publication/language criterias for Article and Category;
 
 
 Version 0.5.0 - 2023/01/16